backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python-headless
pillow
sqlalchemy
passlib[bcrypt]
pyjwt
tensorflow
keras
timm
pydicom